refactor: clean up api_utils ahead of gather-limited extraction
- Delete multitask_gather.py and cache_requests_limited.py: zero
  production callers; both pulled kash settings/shell/web_content
  imports into the otherwise self-contained bundle.
- Move the 8 inline test functions from api_retries.py into
  tests/kash/utils/api_utils/test_api_retries.py (~265 LOC out of the
  production module).
- Mark these prep items done in the utility-extraction plan spec.
